Create BaseDeeplinkDelagate
Update travis

Remove test respository

Fix checkstyle

Fix spacing in BaseDeepLinkDelegate
