+++ deeplinkdispatch/src/main/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+          package="com.airbnb.android.deeplinkdispatch"/>
