+++ deeplinkdispatchbase/src/main/java/com/airbnb/deeplinkdispatch/DeepLinkEntry.java
