+++ deeplinkdispatchbase/src/main/java/com/airbnb/deeplinkdispatch/DeepLinkResult.java
