+++ dependencies.gradle
@@ -1,0 +1,43 @@
+def versions = [
+    kotlinVersion                : '1.2.21',
+    supportLibVersion            : '27.1.1',
+    mockito                      : '1.9.5',
+    testRunner                   : '0.5',
+    espressoIdlingResourceVersion: '2.2.2',
+    roboelectricVersion          : '3.5',
+    spekVersion                  : '1.1.5',
+    androidJunit5Version         : '1.0.12',
+    junitPlatformVersion         : '1.0.1',
+]
+
+ext.versions = versions
+ext.androidConfig = [
+    agpVersion       : '3.1.3',
+    compileSdkVersion: 27,
+    buildToolsVersion: '27.0.3',
+    minSdkVersion    : 16,
+    targetSdkVersion : 27
+]
+
+ext.deps = [
+    autoCommon        : "com.google.auto:auto-common:0.8",
+    androidPlugin     : "com.android.tools.build:gradle:$androidConfig.agpVersion",
+    appCompat         : "com.android.support:appcompat-v7:$versions.supportLibVersion",
+    supportCoreUtils  : "com.android.support:support-core-utils:$versions.supportLibVersion",
+    nexusPlugin       : 'com.bmuschko:gradle-nexus-plugin:2.3.1',
+    appcompat         : "com.android.support:appcompat-v7:$versions.supportLibVersion",
+    kotlinStdLib      : "org.jetbrains.kotlin:kotlin-stdlib:$versions.kotlinVersion",
+    kotlinGradlePlugin: "org.jetbrains.kotlin:kotlin-gradle-plugin:$versions.kotlinVersion",
+    javaPoet          : 'com.squareup:javapoet:1.9.0',
+    jsr305            : 'com.google.code.findbugs:jsr305:3.0.2',
+    okio              : 'com.squareup.okio:okio:1.13.0',
+
+    // Testing
+    junit             : 'junit:junit:4.12',
+    assertJ           : 'org.assertj:assertj-core:1.7.0',
+    roboelectric      : "org.robolectric:robolectric:$versions.roboelectricVersion",
+    mockito           : 'org.mockito:mockito-core:1.10.19',
+    truth             : 'com.google.truth:truth:0.30',
+    compileTesting    : 'com.google.testing.compile:compile-testing:0.10',
+    android           : 'com.google.android:android:4.1.1.4'
+]
