+++ sample/build.gradle
@@ -8,13 +8,13 @@ checkstyle {
 }
 
 android {
-  compileSdkVersion 26
-  buildToolsVersion "26.0.2"
+  compileSdkVersion androidConfig.compileSdkVersion
+  buildToolsVersion androidConfig.buildToolsVersion
 
   defaultConfig {
     applicationId "com.airbnb.deeplinkdispatch.sample"
-    minSdkVersion 16
-    targetSdkVersion 26
+    minSdkVersion androidConfig.minSdkVersion
+    targetSdkVersion androidConfig.compileSdkVersion
     versionCode 1
     versionName "1.0"
   }
@@ -27,15 +27,13 @@ android {
 }
 
 dependencies {
-  compile project(':deeplinkdispatch')
+  implementation project(':deeplinkdispatch')
   annotationProcessor project(':deeplinkdispatch-processor')
+  implementation project(':sample-library')
+  implementation deps.appCompat
 
-  compile project(':sample-library')
-  compile 'com.android.support:appcompat-v7:26.1.0'
-
-  testCompile 'org.robolectric:robolectric:3.5'
-  testCompile 'org.robolectric:shadows-supportv4:3.5'
-  testCompile 'junit:junit:4.12'
+  testImplementation deps.roboelectric
+  testImplementation deps.junit
 }
 
 tasks.withType(JavaCompile) {
